+++ generators/app/templates/gulpfile.js
@@ -1,3 +1,0 @@
-/* eslint no-var: 0 */
-var gulp = require('gulp');
-require('snp-gulp-tasks')(gulp);