+++ generators/app/templates/app/scripts/component/Layout/Layout.js
@@ -1,5 +1,5 @@
 import _Base from '../_Base';
-import './Layout.sass';
+import './Layout.css';
 
 export default _Base.extend({
   regions: {