+++ generators/view/templates/view.js
@@ -1,6 +1,6 @@
 import _Base from '<%= baseClassPath %>';
 import template from './<%= componentName %>.jade';
-import './<%= componentName %>.sass';
+import './<%= componentName %>.css';
 
 export default _Base.extend({
   template,