+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+- Remove snp-gulp-task styles dependency.
+- Remove folder app/styles.
+- Remove font-awesome.
 
 ## [2.0.3] - 2016-02-01
 - Update phantomjs